fixed Taperedcylinder normal caluclation

+++ include/LinearAlgebra/Vector.h
@@ -129,7 +129,7 @@ namespace MRay {
         Vec4(const Vec4 &p1);
 
         //get index
-        float get(int index) const;
+        double get(int index) const;
 
         template<int index>
         void set(float value) {
@@ -144,7 +144,7 @@ namespace MRay {
 
         //Todo: check if this is faster?:
         template<int index>
-        float get() const noexcept {
+        double get() const noexcept {
             assert(index <= 3 & index >= 0);
 #if SET_DATA
             return _mm_cvtss_f32(_mm_shuffle_ps(this->data, this->data, _MM_SHUFFLE(0, 0, 0, index)));
@@ -153,9 +153,9 @@ namespace MRay {
 #endif
         }
 
-        float dot(const Vec4 &rhs) const;
+        double dot(const Vec4 &rhs) const;
 
-        static float dot(const Vec4 &lhs, const Vec4 &rhs);
+        static double dot(const Vec4 &lhs, const Vec4 &rhs);
 
         //compute the cross product as if ti was a Vec3 (only first 3 elements will be used)
         Vec4 cross(const Vec4 &rhs) const;
@@ -163,43 +163,43 @@ namespace MRay {
         static Vec4 cross(const Vec4 &lhs, const Vec4 &rhs);
 
         //cosine of angle between 2 vectors
-        float angle(const Vec4 &rhs) const;
+        double angle(const Vec4 &rhs) const;
 
-        static float angle(const Vec4 &v1, const Vec4 &v2);
+        static double angle(const Vec4 &v1, const Vec4 &v2);
 
-        float abs() const;
+        double abs() const;
 
-        static float abs(const Vec4 &vec);
+        static double abs(const Vec4 &vec);
 
         //sum of all values
-        float sum() const;
+        double sum() const;
 
-        static float sum(const Vec4 &vec);
+        static double sum(const Vec4 &vec);
 
         //sum of first 3 values (sometimes needed when only xyz is required
-        float sum3() const;
+        double sum3() const;
 
-        static float sum3(const Vec4 &vec);
+        static double sum3(const Vec4 &vec);
 
         //max of all values
-        float max() const;
+        double max() const;
 
-        static float max(const Vec4 &vec);
+        static double max(const Vec4 &vec);
         static Vec4 max(Vec4 lhs, const Vec4& rhs);
         //max of first 3 values (sometimes needed when only xyz is required
-        float max3() const;
+        double max3() const;
 
-        static float max3(const Vec4 &vec);
+        static double max3(const Vec4 &vec);
 
         //min of all values
-        float min() const;
+        double min() const;
 
-        static float min(const Vec4 &vec);
+        static double min(const Vec4 &vec);
         static Vec4 min(Vec4 lhs, const Vec4& rhs);
         //min of first 3 values (sometimes needed when only xyz is required
-        float min3() const;
+        double min3() const;
 
-        static float min3(const Vec4 &vec);
+        static double min3(const Vec4 &vec);
 
         //square root elementwise
         static Vec4 sqrt(Vec4 &vec);
@@ -256,7 +256,7 @@ namespace MRay {
 #if SET_DATA
         __m128 data;
 #else
-        float data[4] {0.0,0.0,0.0,0.0};
+        double data[4] {0.0,0.0,0.0,0.0};
 #endif
         friend Matrix4;
         friend Vec2;
